Cienie drzew i optymalizacja

+++ src/engine/lights/Shadow.java
@@ -13,7 +13,7 @@ import engine.utilities.Point;
  */
 public class Shadow {
 
-    public static final int DARK = 0, BRIGHT = 1, BRIGHTEN = 2, DARKEN = 3,
+    public static final short DARK = 0, BRIGHT = 1, BRIGHTEN = 2, DARKEN = 3,
             BRIGHTEN_OBJECT = 4, DARKEN_OBJECT = 5;
     public static final int shadowLength = 32768;
     public final Point point;
